app/requirements.txt
Flask
flask_cors
mysql-connector-python
gunicorn
requests
Flask-SQLAlchemy
Flask-HTTPAuth
passlib
SQLAlchemy
PyMySQL
itsdangerous